requirements.txt
gitpython==3.1.43
googlenews==1.6.15
markupsafe==2.1.5
pyyaml==6.0.2
altair==5.3.0
annotated-types==0.7.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
dateparser==1.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
flatten-json==0.1.13
gitdb==4.0.11
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.4
jieba3k==0.35.1
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
lxml-html-clean==0.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
newspaper3k==0.2.8
nltk==3.8.2
numpy==1.26.1
oauthlib==3.2.2
packaging==24.1
pandas==2.1.1
pillow==10.4.0
pip==23.2.1
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.8.2
pydantic-core==2.20.1
pydeck==0.9.1
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
setuptools==73.0.1
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.36.0
tenacity==8.5.0
tinysegmenter==0.3
tldextract==5.1.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
tweepy==4.14.0
typing-extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.6
watchdog==4.0.1
autocommand==2.2.2
backports.tarfile==1.2.0
importlib-metadata==8.0.0
importlib-resources==6.4.0
inflect==7.3.1
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
more-itertools==10.3.0
platformdirs==4.2.2
tomli==2.0.1
typeguard==4.3.0
wheel==0.43.0
zipp==3.19.2